+++ src/main/java/org/gxfj/iknow/pojo/QuestionState.java
@@ -2,7 +2,7 @@ package org.gxfj.iknow.pojo;
 
 public class QuestionState {
 
-  private long id;
+  private Integer id;
   private String state;
 
 
@@ -10,7 +10,7 @@ public class QuestionState {
     return id;
   }
 
-  public void setId(long id) {
+  public void setId(Integer id) {
     this.id = id;
   }
